+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 ### ✨ Improved
 
 * Update `sdss-clu` to 2.5.3.
+* Remove retries when fetching the buffer is the shape does not match.
+
+### 🔧 Fixed
+
+* Fix pixel shifts in fetched buffers. The shifts were caused by a search-and-replace of the header values in the binary reply, which could remove additional data values when the read pixels matched the header values. The fix removes the header as each binary reply is processed.
+* `ExposureDelegate.reset()` not being called with `await` in `reset` command.
 
 
 ## 0.15.4 - August 10, 2025